Win32 only code cleanup
