+++ source/win32/linux/limits.h
@@ -1,22 +1,17 @@
-// linux/limits.harderr
+// linux/limits.h
+// stab include file to reduce number of #ifdef around #include directives under Windows
+// and to declare missing definitions from Linux's version of limits.h
+#ifndef _WIN32
 
-#ifndef linux_limits_h
-#define linux_limits_h
+#pragma message("this limits.h implementation is for Windows only!")
 
-#include <stdlib.h>
+#else
 
-#define NR_OPEN	        1024
-#define NGROUPS_MAX    65536
-#define ARG_MAX       131072
-#define LINK_MAX         127
-#define MAX_CANON        255
-#define MAX_INPUT        255
-#define NAME_MAX  _MAX_FNAME
+#ifndef __LIMITS_H__
+#define __LIMITS_H__
+
 #define PATH_MAX   _MAX_PATH
-#define PIPE_BUF        4096
-#define XATTR_NAME_MAX   255
-#define XATTR_SIZE_MAX 65536
-#define XATTR_LIST_MAX 65536
-#define RTSIG_MAX	  32
+
+#endif /*__LIMITS_H__*/
 
 #endif
