+++ source/win32/sys/inotify.h
@@ -1,6 +1,14 @@
 // sys/inotify.h
 // stab include to reduce number of #ifdef around #include directives under Windows
 
-#ifndef sys_inotify_h
-#define sys_inotify_h
+#ifndef _WIN32
+
+#pragma message("this inotify.h implementation is for Windows only!")
+
+#else
+
+#ifndef __INOTIFY_H__
+#define __INOTIFY_H__
+#endif /*__INOTIFY_H__*/
+
 #endif
